Remove support for the status extended endpoint
After recieving some feedback from Smartling on https://github.com/Smartling/api-sdk-go/pull/4, it became clear that we did not actually need the extended status endpoint to achieve our goals.  In fact, it was far more efficient to use the regular status endpoint.

To aid with using that endpoint, this commit also adds some helper methods for extracting the FileStatusTranslation for a particular locale